Register RecurlyFormatManager service and use it
The D7 formatting functions were all in recurly.module. This commit
creates a formatting service, RecurlyFormatManager, that houses
the formatting methods. (ex. formatCurrency, formatDate etc.)

The goal is to use dependency injection so that the formatting
service is available in the controllers and forms that need them.

For demonstration purposes, the formatting service has been
injected into the following:

- RecurlyRedeemCouponForm
- RecurlySubscriptionPlansForm
- RecurlySubscriptionListController

The last one is probably the best example, because that one
controller needs to format State, Currency and Date and using the
RecurlyFormatManager service, it can all be done through the one
object.